requirements.txt
Django==6.0.4
djangorestframework==3.17.1
djangorestframework-simplejwt==5.5.1
django-cors-headers==4.7.0
django-filter==25.1
dj-rest-auth==7.2.0
django-allauth==65.15.1
Pillow==11.2.1
reportlab==4.4.0
python-decouple==3.8
requests>=2.32.0
cryptography>=44.0.0
